// app/src/ast/interfaces/LexemeType.ts
export type LexemeType =
      'noun'
    | 'iverb'
    | 'mverb'
    | 'hverb'
    | 'copula'
    | 'then'
    | 'adj'
    | 'existquant'
    | 'uniquant'
    | 'preposition'
    | 'subconj'
    | 'relpron'
    | 'defart'
    | 'indefart'
    | 'fullstop'
    | 'nonsubconj'
    | 'negation'
    | 'contraction'